Feat: System Page and Major Auth Refactor
This commit introduces a new System and Settings page and includes a major refactoring of the authentication system to resolve persistent build failures.

The core of this work was diagnosing and fixing a critical issue where the application would not build for production. The root cause was a version mismatch between the installed `next-auth` library (v4) and the implemented code, which used v5 patterns.

The following changes were made:

1.  **Authentication System Refactoring**:
    -   The entire authentication logic was updated to use NextAuth.js v4 patterns.
    -   All API routes were updated to use `getServerSession(authOptions)` instead of the v5 `auth()` helper.
    -   The API route at `src/app/api/auth/[...nextauth]/route.ts` was corrected to use the v4-style handler initialization.
    -   All related Vitest mocks were updated to correctly mock `getServerSession` and the `authOptions` export, resolving a cascade of test failures.

2.  **New Features**:
    -   Implemented a "System" tab on the settings page for administrators, which displays application version, Node.js version, and database status.
    -   Added a button to the System tab to allow administrators to re-seed the database from the UI.
    -   Created the backing API endpoints (`/api/system/info` and `/api/system/seed`) for this new functionality.
    -   Added a `DepartmentalSpendChart` to the analytics page.

3.  **Bug and Dependency Fixes**:
    -   Installed the `@react-email/render` package, a missing dependency of `resend`, to fix a module resolution error during the build.
    -   Corrected a type incompatibility between the application's `SpendData` type and the `recharts` library by adding an index signature to the type definition, resolving a build-time type error.
