+++ package.json
@@ -16,6 +16,7 @@
     "@auth/prisma-adapter": "^2.10.0",
     "@hookform/resolvers": "^5.2.1",
     "@prisma/client": "^6.16.2",
+    "@react-email/render": "^1.3.1",
     "@tanstack/react-query": "^5.87.1",
     "@tanstack/react-table": "^8.21.3",
     "@tanstack/react-table-devtools": "^8.21.3",
