+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,3 +1,6 @@
-// src/app/api/auth/[...nextauth]/route.ts
-import { handlers } from "@/lib/auth-config";
-export const { GET, POST } = handlers;
+import NextAuth from "next-auth";
+import { authOptions } from "@/lib/auth";
+
+const handler = NextAuth(authOptions);
+
+export { handler as GET, handler as POST };
