+++ src/app/api/dashboard/route.ts
@@ -1,5 +1,6 @@
 import { NextResponse } from "next/server";
-import { auth } from "@/lib/auth-config";
+import { getServerSession } from "next-auth/next";
+import { authOptions } from "@/lib/auth";
 import { prisma } from "@/lib/prisma";
 import { IOM, PurchaseOrder, PaymentRequest } from "@prisma/client";
 import { Kpi } from "@/types/dashboard";
@@ -24,7 +25,7 @@ const formatRecentActivity = (
 
 export async function GET() {
   try {
-    const session = await auth();
+    const session = await getServerSession(authOptions);
     if (!session?.user?.id) {
       return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
     }
