+++ src/app/api/iom/[id]/convert/route.ts
@@ -1,6 +1,7 @@
 // src/app/api/iom/[id]/convert/route.ts
 import { NextRequest, NextResponse } from "next/server";
-import { auth } from "@/lib/auth-config";
+import { getServerSession } from "next-auth/next";
+import { authOptions } from "@/lib/auth";
 import { getIOMById } from "@/lib/iom";
 import { createPurchaseOrder } from "@/lib/po";
 import { prisma } from "@/lib/prisma";
@@ -12,7 +13,7 @@ export async function POST(
   const { id } = await context.params;
 
   try {
-    const session = await auth();
+    const session = await getServerSession(authOptions);
     
     if (!session?.user?.id) {
       return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
