+++ src/app/api/notifications/[id]/route.ts
@@ -1,5 +1,6 @@
 import { NextRequest, NextResponse } from "next/server";
-import { auth } from "@/lib/auth-config";
+import { getServerSession } from "next-auth/next";
+import { authOptions } from "@/lib/auth";
 import { markNotificationAsRead } from "@/lib/notification";
 
 export async function PATCH(
@@ -9,7 +10,7 @@ export async function PATCH(
   const { id } = await context.params;
 
   try {
-    const session = await auth();
+    const session = await getServerSession(authOptions);
     if (!session?.user?.id) {
       return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
     }
