+++ src/app/api/notifications/route.ts
@@ -1,12 +1,13 @@
 // Removed NextRequest from the import statement
 import { NextResponse } from "next/server";
-import { auth } from "@/lib/auth-config";
+import { getServerSession } from "next-auth/next";
+import { authOptions } from "@/lib/auth";
 import { getNotifications } from "@/lib/notification";
 
 // Removed the request parameter from the function signature
 export async function GET() {
   try {
-    const session = await auth();
+    const session = await getServerSession(authOptions);
     if (!session?.user?.id) {
       return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
     }
