+++ src/app/api/permissions/route.test.ts
@@ -3,14 +3,20 @@ import { GET } from './route';
 import { prisma } from '@/lib/prisma';
 import { Session } from 'next-auth';
 import { Permission } from '@prisma/client';
+import { getServerSession } from 'next-auth/next';
+import { authorize } from '@/lib/auth-utils';
 
 vi.mock('@/lib/prisma');
-vi.mock('@/lib/auth-config', () => ({
-  auth: vi.fn(),
+vi.mock('next-auth/next', () => ({
+  getServerSession: vi.fn(),
+}));
+vi.mock('@/lib/auth', () => ({
+  authOptions: {},
 }));
 vi.mock('@/lib/auth-utils', () => ({
   authorize: vi.fn(),
 }));
+
 
 const mockSession: Session = {
   user: { id: '1', name: 'Test User', email: 'test@example.com' },
@@ -23,11 +29,9 @@ describe('GET /api/permissions', () => {
   });
 
   it('should return a list of permissions successfully', async () => {
-    const { auth } = await import('@/lib/auth-config');
-    const { authorize } = await import('@/lib/auth-utils');
-    vi.mocked(auth).mockResolvedValue(mockSession);
+    vi.mocked(getServerSession).mockResolvedValue(mockSession);
     vi.mocked(authorize).mockReturnValue(true);
-    const mockPermissions: Permission[] = [{ id: '1', name: 'TEST_PERMISSION' }];
+    const mockPermissions: Permission[] = [{ id: '1', name: 'TEST_PERMISSION', description: 'A test permission' }];
     vi.mocked(prisma.permission.findMany).mockResolvedValue(mockPermissions);
 
     const response = await GET();
@@ -39,17 +43,14 @@ describe('GET /api/permissions', () => {
   });
 
   it('should return 401 for unauthenticated users', async () => {
-    const { auth } = await import('@/lib/auth-config');
-    vi.mocked(auth).mockResolvedValue(null);
+    vi.mocked(getServerSession).mockResolvedValue(null);
 
     const response = await GET();
     expect(response.status).toBe(401);
   });
 
   it('should return 403 for unauthorized users', async () => {
-    const { auth } = await import('@/lib/auth-config');
-    const { authorize } = await import('@/lib/auth-utils');
-    vi.mocked(auth).mockResolvedValue(mockSession);
+    vi.mocked(getServerSession).mockResolvedValue(mockSession);
     vi.mocked(authorize).mockImplementation(() => {
       throw new Error('Not authorized');
     });
