+++ src/app/api/pr/[id]/route.ts
@@ -1,7 +1,8 @@
 // src/app/api/pr/[id]/route.ts
 
 import { NextRequest, NextResponse } from "next/server";
-import { auth } from "@/lib/auth-config";
+import { getServerSession } from "next-auth/next";
+import { authOptions } from "@/lib/auth";
 import { authorize } from "@/lib/auth-utils"; // Make sure to import authorize
 import { getPRById, updatePRStatus } from "@/lib/pr";
 
@@ -12,7 +13,7 @@ export async function GET(
   const { id } = await context.params;
 
   try {
-    const session = await auth();
+    const session = await getServerSession(authOptions);
 
     if (!session?.user?.id) {
       return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
@@ -47,7 +48,7 @@ export async function PATCH(
   const { id } = await context.params;
 
   try {
-    const session = await auth();
+    const session = await getServerSession(authOptions);
 
     if (!session?.user?.id) {
       return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
