+++ src/app/api/pusher/auth/route.ts
@@ -1,9 +1,10 @@
 import { pusherServer } from '@/lib/pusher';
-import { auth } from "@/lib/auth-config";
+import { getServerSession } from "next-auth/next";
+import { authOptions } from "@/lib/auth";
 import { NextRequest, NextResponse } from 'next/server';
 
 export async function POST(req: NextRequest) {
-  const session = await auth();
+  const session = await getServerSession(authOptions);
 
   if (!session?.user?.id) {
     return NextResponse.json({ error: 'Forbidden' }, { status: 403 });
