+++ src/app/api/roles/[id]/route.ts
@@ -1,5 +1,6 @@
 import { NextRequest, NextResponse } from 'next/server';
-import { auth } from "@/lib/auth-config";
+import { getServerSession } from 'next-auth/next';
+import { authOptions } from "@/lib/auth";
 import { authorize } from '@/lib/auth-utils';
 import { prisma } from '@/lib/prisma';
 
@@ -11,7 +12,7 @@ export async function GET(
 ) {
   try {
     const params = await paramsPromise;
-    const session = await auth();
+    const session = await getServerSession(authOptions);
     if (!session?.user) {
       return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
     }
@@ -52,7 +53,7 @@ export async function PUT(
 ) {
     try {
       const params = await paramsPromise;
-      const session = await auth();
+      const session = await getServerSession(authOptions);
       if (!session?.user) {
         return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
       }
