+++ src/app/api/settings/[key]/route.ts
@@ -1,6 +1,7 @@
 import { NextRequest, NextResponse } from 'next/server';
 import { prisma } from '@/lib/prisma';
-import { auth } from '@/lib/auth-config';
+import { getServerSession } from 'next-auth/next';
+import { authOptions } from '@/lib/auth';
 import { authorize } from '@/lib/auth-utils';
 
 export async function GET(
@@ -8,7 +9,7 @@ export async function GET(
   context: { params: Promise<{ key: string }> }
 ) {
   try {
-    const session = await auth();
+    const session = await getServerSession(authOptions);
     if (!session) {
       return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
     }
@@ -43,7 +44,7 @@ export async function PUT(
   context: { params: Promise<{ key: string }> }
 ) {
   try {
-    const session = await auth();
+    const session = await getServerSession(authOptions);
     if (!session) {
       return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
     }
