+++ src/app/api/settings/route.ts
@@ -1,11 +1,12 @@
 import { NextResponse } from 'next/server';
 import { prisma } from '@/lib/prisma';
-import { auth } from '@/lib/auth-config';
+import { getServerSession } from 'next-auth/next';
+import { authOptions } from '@/lib/auth';
 import { authorize } from '@/lib/auth-utils';
 
 export async function GET() {
   try {
-    const session = await auth();
+    const session = await getServerSession(authOptions);
     if (!session) {
       return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
     }
