+++ src/lib/auth.ts
@@ -1,9 +1,91 @@
-// src/lib/auth.ts
-// For server-side auth, we need to import from next-auth/next
-// For client-side, we'll import from next-auth/react
+import { type AuthOptions } from "next-auth";
+import Credentials from "next-auth/providers/credentials";
+import { prisma } from "./prisma";
+import bcrypt from "bcryptjs";
+import { z } from "zod";
 
-// Export client-side functions
-export { signIn, signOut } from "next-auth/react";
+const credentialsSchema = z.object({
+  email: z.string().email(),
+  password: z.string().min(1),
+});
 
-// For server-side session handling, we'll create a separate utility
-// or import directly where needed
+export const authOptions: AuthOptions = {
+  session: { strategy: "jwt" },
+  secret: process.env.NEXTAUTH_SECRET,
+  providers: [
+    Credentials({
+      credentials: {
+        email: {},
+        password: {},
+      },
+      authorize: async (credentials) => {
+        const { email, password } = credentialsSchema.parse(credentials);
+
+        const user = await prisma.user.findUnique({
+          where: { email },
+        });
+
+        if (!user) {
+          return null;
+        }
+
+        const passwordValid = await bcrypt.compare(
+          password,
+          user.password || ""
+        );
+
+        if (!passwordValid) {
+          return null;
+        }
+
+        return {
+          id: user.id,
+        };
+      },
+    }),
+  ],
+  callbacks: {
+    async jwt({ token, user }) {
+      if (user) {
+        token.id = user.id;
+      }
+
+      const dbUser = await prisma.user.findUnique({
+        where: { id: token.id as string },
+        include: {
+          role: {
+            include: { permissions: { include: { permission: true } } },
+          },
+        },
+      });
+
+      if (dbUser) {
+        token.name = dbUser.name;
+        token.email = dbUser.email;
+        token.roleId = dbUser.role?.id;
+        token.roleName = dbUser.role?.name;
+        token.permissions = dbUser.role?.permissions.map(
+          (p) => p.permission.name
+        );
+      }
+
+      return token;
+    },
+    async session({ session, token }) {
+      if (session.user) {
+        session.user.id = token.id as string;
+        session.user.name = token.name as string;
+        session.user.email = token.email as string;
+        session.user.role = {
+          id: token.roleId as string,
+          name: token.roleName as string
+        };
+        session.user.permissions = token.permissions as string[];
+      }
+      return session;
+    },
+  },
+  pages: {
+    signIn: "/login",
+  },
+};
