+++ src/types/analytics.d.ts
@@ -2,6 +2,8 @@ export interface SpendData {
   name: string;
   Total?: number;
   value?: number;
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  [key: string]: any;
 }
 
 export interface AnalyticsData {
